pyproject.toml
[project]
name = "evalium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.0.300",
    "langgraph>=0.1.0",
    "numpy>=1.21.0",
    "openai>=0.27.0",
    "openpyxl>=3.0.0",
    "pandas>=1.3.0",
    "python-dotenv>=0.21.0",
    "scikit-learn>=1.0.0",
]
